+++ R/ictr.R
@@ -3,8 +3,8 @@
 #' ```{r child=extdata_path("child/intro-ictr.Rmd")}
 #' ```
 #'
-#' @inheritParams ictr_score_inputs
-#' @inheritParams ictr_score_companies
+#' @inheritParams ictr_at_product_level
+#' @inheritParams ictr_at_company_level
 #'
 #' @family ICTR functions
 #'
@@ -14,13 +14,13 @@
 #'
 #' @examples
 #' ictr(ictr_companies, ictr_inputs)
-ictr <- function(companies, inputs, low_threshold = 0.3, high_threshold = 0.7) {
-  out <- inputs |>
-    ictr_score_inputs(
+ictr <- function(companies, co2, low_threshold = 0.3, high_threshold = 0.7) {
+  out <- co2 |>
+    ictr_at_product_level(
       low_threshold = low_threshold,
       high_threshold = high_threshold
     ) |>
-    ictr_score_companies(companies)
+    ictr_at_company_level(companies)
 
   out |>
     xctr_rename() |>