+++ NEWS.md
@@ -1,5 +1,13 @@
 <!-- NEWS.md is maintained by https://cynkra.github.io/fledge, do not edit -->
 
+# tiltIndicator 0.0.0.9029
+
+* In PCTR, if a company matches no input, all shares are now `NA` (#205).
+
+# tiltIndicator 0.0.0.9028
+
+* In ICTR, if a company matches no input, all shares are now `NA` (@kalashsinghal #202).
+
 # tiltIndicator 0.0.0.9027
 
 * Even if companies has a *uuid absent in inputs/co2, all shares now sum 1  (@kalashsinghal #197).